Refactor sc-memory
